requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12,<3.14
requests>=2.25.0,<3.0
